server/requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.5.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=2.3.0
TTS>=0.15.0
accelerate>=0.20.0
sentencepiece>=0.1.99 